% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePredTree.R
\name{ePredTree}
\alias{ePredTree}
\title{Predict responses through an explainable RF}
\usage{
ePredTree(info, data, target = "1")
}
\arguments{
\item{info}{bla bla bla}

\item{data}{bla bla bla}

\item{target}{bla bla bla}
}
\value{
an object.
}
\description{
It predicts classification tree responses
}
\examples{

}
