% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etree-package.R
\docType{package}
\name{etree-package}
\alias{etree-package}
\title{etree: Classification and Regression With Structured and Mixed-Type Data}
\description{
Implementation of Energy Trees, a statistical model to perform classification
and regression with structured and mixed-type data. This is made possible by
starting from the statistically sound structure of Conditional Trees, and
using Energy Statistics to test independence between data objects of possibly
structured and possibly different nature. The type of structured objects
covered so far are functions and graphs. The package builds upon 'partykit'
to provide functionalities for fitting, printing, plotting, and predicting
with Energy Trees.
}
