% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{etree-size}
\alias{etree-size}
\alias{depth}
\alias{width}
\title{Size of Energy Trees}
\usage{
depth(x, ...)
}
\arguments{
\item{x}{An Energy Tree object.}

\item{...}{Additional arguments.}
}
\description{
Depth and width of an Energy Tree.
}
\section{Functions}{
\itemize{
\item \code{depth}: Depth of the three.
}}

