% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eTree.R
\name{eTree}
\alias{eTree}
\title{Explainable Random Forest}
\usage{
eTree(
  Y,
  X,
  response,
  setting = list(impTotal = 0.1, maxDec = 0.01, n = 5, level = 5, tMax = 5)
)
}
\arguments{
\item{Y}{bla bla bla}

\item{X}{bla bla bla}

\item{response}{bla bla bla}

\item{setting}{bla bla bla}
}
\value{
an object.
}
\description{
It creates an explainable tree for Random Forest
}
\examples{

}
