% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Roc curve}
\usage{
roc(response, scores, target = "1")
}
\arguments{
\item{response}{bla bla bla}

\item{scores}{bla bla bla}

\item{target}{bla bla bla}
}
\value{
an object.
}
\description{
It calculates roc curve
}
\examples{

}
