% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp.R
\name{vimp}
\alias{vimp}
\title{Variable Importance}
\usage{
vimp(info, response, X, graph = TRUE)
}
\arguments{
\item{info}{bla bla bla}

\item{response}{bla bla bla}

\item{X}{bla bla bla}

\item{graph}{bla bla bla}
}
\value{
an object.
}
\description{
It calculate variable importance of an explainable tree
}
\examples{

}
