% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etree-package.R
\docType{package}
\name{etree-package}
\alias{etree-package}
\title{etree: Classification and Regression With Structured and Mixed-Type Data}
\description{
Implementation of Energy Trees, a statistical model to perform classification
and regression with structured and mixed-type data. The model has a similar
structure to Conditional Trees, but brings in Energy Statistics to test
independence between variables that are possibly structured and of different
nature. Currently, the package covers functions and graphs as structured
covariates. It builds upon 'partykit' to provide functionalities for fitting,
printing, plotting, and predicting with Energy Trees.
}
