% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{etree-size}
\alias{etree-size}
\alias{depth}
\alias{width}
\title{Size of Energy Trees}
\usage{
depth(x, ...)

width(x, ...)
}
\arguments{
\item{x}{An object of class \code{etree}.}

\item{...}{Additional arguments.}
}
\value{
\code{depth()} returns the depth of the tree and \code{width()} gives the
number of terminal nodes.
}
\description{
Depth and width of an Energy Tree.
}
\section{Functions}{
\itemize{
\item \code{depth}: Depth of the three.

\item \code{width}: Number of terminal nodes in the tree.
}}

