% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etree}
\alias{etree}
\title{Energy Tree}
\usage{
etree(
  response,
  covariates,
  case.weights = NULL,
  minbucket = 1,
  alpha = 0.05,
  R = 1000,
  split.type = "coeff",
  coef.split.type = "test",
  nb = 5
)
}
\arguments{
\item{response}{response variable (either numeric or factor).}

\item{covariates}{covariates. Must be provided as a list, where each element of the list is a different variable.}

\item{case.weights}{an optional numeric vector of weights to be used in the fitting process.}

\item{minbucket}{minimum number of observations that each terminal node must contain. Default is 1.}

\item{alpha}{significance level for the global test of association and, if \code{split.type = "coeff"} and \code{coef.split.type = "test"}, for the test used in each split. Default is 0.05.}

\item{R}{number of replicates for the global test of association and, if \code{split.type = "coeff"} and \code{coef.split.type = "test"}, for the test used in each split. Deafult is 1000.}

\item{split.type}{type of the split when covariates are "complex" (i.e. they are not numeric or factor). This variable can be se to either "coeff" (only available when there is a coeffient representation) or "cluster"}

\item{coef.split.type}{either "variance" or "test"}

\item{nb}{number of basis to use for fdata covariates if \code{split.type = "coeff"}.}
}
\description{
Fits an energy tree for classification/regression using mixed type data.
}
\examples{
add_numbers(1, 2) ## returns 3

}
