% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.party}
\alias{predict.party}
\title{Energy Tree Predictions}
\usage{
\method{predict}{party}(object, newdata = NULL, perm = NULL, ...)
}
\arguments{
\item{object}{object of class party.}

\item{newdata}{an optional list of variables used to make predictions. If omitted, the fitted values are used.}

\item{perm}{an optional character vector of variable names. Splits of nodes with a primary split in any of these variables will be permuted (after dealing with surrogates). Note that surrogate split in the \code{perm} variables will no be permuted.}

\item{...}{additional arguments.}
}
\description{
Compute predictions based on an Energy Tree Fit.
}
\examples{
add_numbers(1, 2) ## returns 3

}
